Conflicts:
	modules/photoeffects/doc/edge_blur.rst
	modules/photoeffects/doc/fade_color.rst
	modules/photoeffects/doc/film_grain.rst
	modules/photoeffects/doc/sepia.rst
	modules/photoeffects/doc/tint.rst
	modules/photoeffects/doc/vignette.rst
	modules/photoeffects/doc/warmify.rst
	modules/photoeffects/include/opencv2/photoeffects.hpp
	modules/photoeffects/samples/antique_sample.cpp
	modules/photoeffects/samples/edgeBlur_sample.cpp
	modules/photoeffects/samples/fadeColor_sample.cpp
	modules/photoeffects/samples/matte_sample.cpp
	modules/photoeffects/samples/sepia_sample.cpp
	modules/photoeffects/samples/tint_sample.cpp
	modules/photoeffects/samples/vignette_sample.cpp
	modules/photoeffects/src/antique.cpp
	modules/photoeffects/src/boostColor.cpp
	modules/photoeffects/src/edgeBlur.cpp
	modules/photoeffects/src/fadeColor.cpp
	modules/photoeffects/src/filmGrain.cpp
	modules/photoeffects/src/glow.cpp
	modules/photoeffects/src/matte.cpp
	modules/photoeffects/src/sepia.cpp
	modules/photoeffects/src/tint.cpp
	modules/photoeffects/src/vignette.cpp
	modules/photoeffects/src/warmify.cpp
	modules/photoeffects/test/antique_test.cpp
	modules/photoeffects/test/edgeBlur_test.cpp
	modules/photoeffects/test/fadeColor_test.cpp
	modules/photoeffects/test/filmGrain_test.cpp
	modules/photoeffects/test/glow_test.cpp
	modules/photoeffects/test/matte_test.cpp
	modules/photoeffects/test/sepia_test.cpp
	modules/photoeffects/test/tint_test.cpp
	modules/photoeffects/test/vignette_test.cpp
	modules/photoeffects/test/warmify_test.cpp

+++ modules/photoeffects/samples/antique_sample.cpp
@@ -15,8 +15,8 @@ const char* helper = "./antique_sample <src> <texture> <alpha>\n\
 \t<texture> - file name contained the texture image, must be 3-channel RGB-image\n\
 \t<alpha> - float coefficient of intensity texture applying, must be real between 0 to 1\n";
 
-int processArguments(int argc, char** argv, Mat &src,
-    Mat &texture, float &alpha);
+
+int processArguments(int argc, char** argv, Mat &src, Mat &texture, float &alpha);
 
 int main(int argc, char** argv)
 {
@@ -40,8 +40,8 @@ int main(int argc, char** argv)
     return 0;
 }
 
-int processArguments(int argc, char** argv, Mat &src,
-    Mat &texture, float &alpha)
+int processArguments(int argc, char** argv, Mat &src, Mat &texture, float &alpha)
+
 {
     if(argc < 4)
     {
@@ -49,6 +49,6 @@ int processArguments(int argc, char** argv, Mat &src,
     }
     src = imread(argv[1], CV_LOAD_IMAGE_COLOR);
     texture = imread(argv[2], CV_LOAD_IMAGE_COLOR);
-    alpha = atof(argv[3]);
+    alpha = (float)atof(argv[3]);
     return 0;
 }
